Define private non-member function readRank.
Defined readRank, a private function which is only meant to be used by
one of the readFEN overloads. The functions parses a single rank from
the first field in a FEN record. It checks that the record does not
contain any illegal characters, that the rank is not empty, that the
rank does not contain more than 8 chars, and that the sum of the empty
squres does not exceed the maximum length of a valid rank.