Define readEnPassant(const string&).
The function reads the 4th field in a FEN notation record. The function
tests that the column letter is valid, and that the row number is valid.
If the field contains something (i.e. is not a dash), then it returns a
PieceMove that represents the last move as a pawn move of two squares
from either the 2nd rank to the 4th rank, or from the 7th rank to the
5th rank. If its a dash, then it returns a PieceMove with blank squares.