Pur all return type on seperate line
Put all return types, and specifiers that precede the function name, on
a separate line. This makes code more readable for multiple reasons:
more emphasis placed on return type because it is on its own line; more
emphasis is placed on function name, because it is the first thing that
appears on a line; decreases number of instances where parameters have
to be split across more than one line because the function name is too
long.