Define move assignment for BasicBoard
Move operation is handled differently in move assignment vs. move
construction. The main difference is that in assignment this has already
been constructed, thus has a valid mArr pointer, so we can swap data,
but in move construction this hasn't been initialized, and it is cheaper
to simply transfer ownership of data than to allocate and then swap.